requirements.txt
Flask==3.0.0
rapidfuzz==3.9.6
regex==2024.7.24
python-docx==1.1.2
dateparser==1.2.0
openai>=1.40.0
tenacity>=8.2.3
pdfplumber>=0.11.4
pypdf>=5.0.1
gunicorn==21.2.0
